// frontend/src/shared/api/generated/model/lessonResultRead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */
import type { LessonResultBreakdown } from './lessonResultBreakdown';

export interface LessonResultRead {
  earned_marks: number;
  total_marks: number;
  percent: number;
  stars: number;
  breakdown: LessonResultBreakdown[];
  need_review: string[];
}